implement runComposeUiTest
