+++ compose/ui/ui-test-junit4/src/jvmMain/kotlin/androidx/compose/ui/test/junit4/UncaughtExceptionHandler.jvm.kt
@@ -16,4 +16,4 @@
 
 package androidx.compose.ui.test.junit4
 
-internal actual fun synchronized(lock: Any, block: () -> Unit) = kotlin.synchronized(lock, block)
+internal actual inline fun <T> synchronized(lock: Any, block: () -> T): T = kotlin.synchronized(lock, block)
